requirements.txt
selenium
schedule
pytesseract
Pillow